refactored settings midware into component parts, don't try that again...
