+++ client/src/fleet/settings/cost-centre-group/index.js
@@ -1,5 +1,7 @@
+export { costCentreGroupList } from './cost-centre-group-list';
 export { costCentreGroupActions } from './cost-centre-group-actions';
 export { costCentreGroupReducer } from './cost-centre-group-reducer';
+export { costCentreGroupSagas } from './cost-centre-group-sagas';
 export {
   getCostCentreGroupList,
   getCostCentreGroups,
