+++ client/src/fleet/settings/cost-centre/index.js
@@ -1,5 +1,8 @@
+export { costCentreList } from './cost-centre-list';
 export { costCentreActions } from './cost-centre-actions';
 export { costCentreReducer } from './cost-centre-reducer';
+export { costCentreSagas } from './cost-centre-sagas';
+
 export {
   getCostCentres,
   getCostCentreById,
