+++ client/src/fleet/settings/fims-period/index.js
@@ -1,5 +1,7 @@
+export { fimsPeriodList } from './fims-period-list';
 export { fimsPeriodActions } from './fims-period-actions';
 export { fimsPeriodReducer } from './fims-period-reducer';
+export { fimsPeriodSagas } from './fims-period-sagas';
 export {
   getFimsPeriodById,
   getFimsPeriodList,
