+++ client/src/fleet/settings/setting-sagas.js
@@ -1,218 +1,13 @@
-import { call, fork, put, take, takeEvery } from 'redux-saga/effects';
-import { authActions } from 'src/common/auth/index';
-import { settingActions } from './setting-actions';
-import { costCentreList } from './cost-centre/cost-centre-list';
-import { costCentreGroupList } from './cost-centre-group/cost-centre-group-list';
-import { transactionTypeList } from './transaction-type/transaction-type-list';
-import { fimsPeriodList } from './fims-period/fims-period-list';
-
-function* importFimsPeriod({ payload }) {
-  //console.log(payload);
-  const fimsPeriod = yield call(
-    [fimsPeriodList, fimsPeriodList.importFimsPeriod],
-    payload.id
-  );
-  yield put(settingActions.importFimsPeriodFulfilled(fimsPeriod));
-}
-
-function* removeFimsPeriod({ payload }) {
-  const id = yield call([fimsPeriodList, fimsPeriodList.remove], payload.id);
-  yield put(settingActions.removeFimsPeriodFulfilled(id));
-}
-
-function* loadFimsPeriods() {
-  const fimsPeriods = yield call([fimsPeriodList, fimsPeriodList.list]);
-  yield put(settingActions.loadFimsPeriodsFulfilled(fimsPeriods));
-}
-
-function* postFimsBatch({ payload }) {
-  const { fimsBatch } = payload;
-  const postFimsBatchResult = yield call(
-    [fimsPeriodList, fimsPeriodList.postFimsBatch],
-    { fimsBatch }
-  );
-  yield put(settingActions.postFimsBatchFulfilled(postFimsBatchResult));
-}
-
-function* removeCostCentre({ payload }) {
-  let result = yield call(
-    [costCentreList, costCentreList.remove],
-    payload.costCentre.uid
-  );
-  yield put(settingActions.removeCostCentreFulfilled(result));
-}
-
-function* updateCostCentre({ payload }) {
-  let result = yield call([costCentreList, costCentreList.update], payload.id, {
-    costCentre: payload.costCentre,
-    changes: payload.changes
-  });
-  yield put(settingActions.updateCostCentreFulfilled(result.costCentre));
-}
-
-function* loadAllCostCentreGroups() {
-  const costCentreGroups = yield call([
-    costCentreGroupList,
-    costCentreGroupList.list
-  ]);
-  yield put(settingActions.loadCostCentreGroupsFulfilled(costCentreGroups));
-}
-
-function* loadAllTransactionTypes() {
-  const transactionTypes = yield call([
-    transactionTypeList,
-    transactionTypeList.list
-  ]);
-  yield put(settingActions.loadTransactionTypesFulfilled(transactionTypes));
-}
-
-function* createCostCentreGroup({ payload: { costCentreGroup } }) {
-  let result = yield call([costCentreGroupList, costCentreGroupList.insert], {
-    costCentreGroup
-  });
-  yield put(
-    settingActions.createCostCentreGroupFulfilled(result.costCentreGroup)
-  );
-}
-
-function* updateCostCentreGroup({ payload: { id, changes } }) {
-  let result = yield call(
-    [costCentreGroupList, costCentreGroupList.update],
-    id,
-    { changes }
-  );
-  yield put(
-    settingActions.updateCostCentreGroupFulfilled(result.costCentreGroup)
-  );
-}
-
-function* removeCostCentreGroup({ payload: { costCentreGroup } }) {
-  let result = yield call(
-    [costCentreGroupList, costCentreGroupList.remove],
-    costCentreGroup.id
-  );
-  if (result.status === 'deleted') {
-    yield put(settingActions.removeCostCentreGroupFulfilled(costCentreGroup));
-  } else {
-    yield put(settingActions.removeCostCentreGroupFailed(result));
-  }
-}
-
-//=====================================
-//  WATCHERS
-//-------------------------------------
-
-function* watchAuthentication() {
-  while (true) {
-    let { payload } = yield take(authActions.SIGN_IN_FULFILLED);
-    costCentreList.token = payload.idToken;
-    costCentreGroupList.token = payload.idToken;
-    transactionTypeList.token = payload.idToken;
-    fimsPeriodList.token = payload.idToken;
-    yield take([authActions.SIGN_OUT_FULFILLED]);
-    costCentreList.token = null;
-    costCentreGroupList.token = payload.null;
-    transactionTypeList.token = null;
-    fimsPeriodList.token = null;
-  }
-}
-
-function* watchIdTokenRefresh() {
-  while (true) {
-    const { payload } = yield take(authActions.REFRESH_ID_TOKEN_FULFILLED);
-    costCentreList.token = payload.idToken;
-    costCentreGroupList.token = payload.idToken;
-    transactionTypeList.token = payload.idToken;
-    fimsPeriodList.token = payload.idToken;
-  }
-}
-
-function* loadAllCostCentres() {
-  const costCentres = yield call([costCentreList, costCentreList.list]);
-  yield put(settingActions.loadCostCentresFulfilled(costCentres));
-}
-
-function* watchLoadCostCentres() {
-  yield takeEvery(settingActions.LOAD_COST_CENTRES, loadAllCostCentres);
-}
-
-function* watchRemoveCostCentre() {
-  yield takeEvery(settingActions.REMOVE_COST_CENTRE, removeCostCentre);
-}
-
-function* watchUpdateCostCentre() {
-  yield takeEvery(settingActions.UPDATE_COST_CENTRE, updateCostCentre);
-}
-
-function* watchLoadCostCentreGroups() {
-  yield takeEvery(
-    settingActions.LOAD_COST_CENTRE_GROUPS,
-    loadAllCostCentreGroups
-  );
-}
-
-function* watchCreateCostCentreGroup() {
-  yield takeEvery(
-    settingActions.CREATE_COST_CENTRE_GROUP,
-    createCostCentreGroup
-  );
-}
-
-function* watchUpdateCostCentreGroup() {
-  yield takeEvery(
-    settingActions.UPDATE_COST_CENTRE_GROUP,
-    updateCostCentreGroup
-  );
-}
-
-function* watchRemoveCostCentreGroup() {
-  yield takeEvery(
-    settingActions.REMOVE_COST_CENTRE_GROUP,
-    removeCostCentreGroup
-  );
-}
-
-function* watchLoadTransactionTypes() {
-  yield takeEvery(
-    settingActions.LOAD_TRANSACTION_TYPES,
-    loadAllTransactionTypes
-  );
-}
-
-function* watchFimsLoadPeriods() {
-  yield takeEvery(settingActions.LOAD_FIRMS_PERIODS, loadFimsPeriods);
-}
-
-function* watchPostFimsBatch() {
-  yield takeEvery(settingActions.POST_FIMS_BATCH, postFimsBatch);
-}
-
-function* watchRemoveFimsPeriod() {
-  yield takeEvery(settingActions.REMOVE_FIRMS_PERIOD, removeFimsPeriod);
-}
-
-function* watchImportFimsPeriod() {
-  yield takeEvery(settingActions.IMPORT_FIRMS_PERIOD, importFimsPeriod);
-}
-
-//=====================================
-//  COST_CENTRE SAGAS
-//-------------------------------------
+import {
+  costCentreSagas,
+  costCentreGroupSagas,
+  fimsPeriodSagas,
+  transactionTypeSagas
+} from './';
 
 export const settingSagas = [
-  fork(watchAuthentication),
-  fork(watchIdTokenRefresh),
-  fork(watchLoadCostCentres),
-  //fork(watchCreateCostCentre),
-  fork(watchRemoveCostCentre),
-  fork(watchUpdateCostCentre),
-  fork(watchLoadCostCentreGroups),
-  fork(watchLoadTransactionTypes),
-  fork(watchFimsLoadPeriods),
-  fork(watchPostFimsBatch),
-  fork(watchRemoveFimsPeriod),
-  fork(watchImportFimsPeriod),
-  fork(watchUpdateCostCentreGroup),
-  fork(watchCreateCostCentreGroup),
-  fork(watchRemoveCostCentreGroup)
+  ...costCentreSagas,
+  ...costCentreGroupSagas,
+  ...fimsPeriodSagas,
+  ...transactionTypeSagas
 ];
