+++ client/src/fleet/settings/transaction-type/index.js
@@ -1,5 +1,7 @@
+export { transactionTypeList } from './transaction-type-list';
 export { transactionTypeActions } from './transaction-type-actions';
 export { transactionTypeReducer } from './transaction-type-reducer';
+export { transactionTypeSagas } from './transaction-type-sagas';
 export {
   getTransactionTypeById,
   getTransactionTypeList,
